+++ coupon-generator-backend/src/main/java/org/example/service/impl/UserServiceImpl.java
@@ -17,6 +17,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserServiceImpl implements UserService {
@@ -51,7 +52,7 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public ApiResponse getUsers() {
-        logger.info("Get Users");
+        logger.info("Get Users starts");
         List<UserEntity> userEntities = userRepository.findAll();
 
         ApiResponse apiResponse = new ApiResponse();
@@ -63,32 +64,69 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public ApiResponse deleteUser(int userId) {
-        logger.info("Delete User Starts");
-        userRepository.deleteById(userId);
+    public ApiResponse getUserById(int userId) {
+        logger.info("Get UserById starts");
+
+        Optional<UserEntity> optionalUserEntity = userRepository.findById(userId);
 
         ApiResponse apiResponse = new ApiResponse();
-        apiResponse.setResponseCode(ResponseCodes.SUCCESS);
-        apiResponse.setStatusCode(RequestStatus.SUCCESS.getStatusCode());
-        apiResponse.setMessage(messages.getMessageForResponseCode(ResponseCodes.USER_DELETED, null));
-        logger.info("User Deleted {} " , userId);
+        if(optionalUserEntity.isPresent()){
+            UserEntity userEntity = optionalUserEntity.get();
+            apiResponse.setUserEntity(userEntity);
+            apiResponse.setResponseCode(ResponseCodes.SUCCESS);
+            apiResponse.setStatusCode(RequestStatus.SUCCESS.getStatusCode());
+        }else {
+            apiResponse.setResponseCode(ResponseCodes.NOT_FOUND);
+            apiResponse.setStatusCode(RequestStatus.NOT_FOUND.getStatusCode());
+            apiResponse.setMessage(messages.getMessageForResponseCode(ResponseCodes.USER_NOT_FOUND, null));
+        }
+        logger.info("Get UserById ends");
+        return apiResponse;
+    }
+
+    @Override
+    public ApiResponse deleteUser(int userId) {
+        logger.info("Delete User starts");
+        Optional<UserEntity> optionalUserEntity = userRepository.findById(userId);
+        ApiResponse apiResponse = new ApiResponse();
+
+        if(optionalUserEntity.isPresent()){
+            userRepository.deleteById(userId);
+            apiResponse.setResponseCode(ResponseCodes.SUCCESS);
+            apiResponse.setStatusCode(RequestStatus.SUCCESS.getStatusCode());
+            apiResponse.setMessage(messages.getMessageForResponseCode(ResponseCodes.USER_DELETED, null));
+            logger.info("User Deleted {} " , userId);
+        }else{
+            apiResponse.setResponseCode(ResponseCodes.NOT_FOUND);
+            apiResponse.setStatusCode(RequestStatus.NOT_FOUND.getStatusCode());
+            apiResponse.setMessage(messages.getMessageForResponseCode(ResponseCodes.USER_NOT_FOUND, null));
+        }
+        logger.info("Delete User ends");
         return apiResponse;
     }
 
     @Override
     public ApiResponse updateUser(UserDTO userDTO, int userId) {
-        logger.info("Update User Starts");
-        UserEntity userEntity = userRepository.findById(userId).get();
-
-        userEntity.setUserName(userDTO.getUserName());
-        userEntity.setAddress(userDTO.getAddress());
-        userRepository.save(userEntity);
-
+        logger.info("Update User starts");
+        Optional<UserEntity> optionalUserEntity = userRepository.findById(userId);
         ApiResponse apiResponse = new ApiResponse();
-        apiResponse.setResponseCode(ResponseCodes.SUCCESS);
-        apiResponse.setStatusCode(RequestStatus.SUCCESS.getStatusCode());
-        apiResponse.setMessage(messages.getMessageForResponseCode(ResponseCodes.USER_UPDATED, null));
-        logger.info("User Updated {} " , userEntity.getUserId());
+
+        if(optionalUserEntity.isPresent()){
+            UserEntity userEntity = optionalUserEntity.get();
+            userEntity.setUserName(userDTO.getUserName());
+            userEntity.setAddress(userDTO.getAddress());
+            userRepository.save(userEntity);
+            apiResponse.setResponseCode(ResponseCodes.SUCCESS);
+            apiResponse.setStatusCode(RequestStatus.SUCCESS.getStatusCode());
+            apiResponse.setMessage(messages.getMessageForResponseCode(ResponseCodes.USER_UPDATED, null));
+            logger.info("User Updated {} " , userEntity.getUserId());
+        }else{
+            apiResponse.setResponseCode(ResponseCodes.NOT_FOUND);
+            apiResponse.setStatusCode(RequestStatus.NOT_FOUND.getStatusCode());
+            apiResponse.setMessage(messages.getMessageForResponseCode(ResponseCodes.USER_NOT_FOUND, null));
+            logger.info("User Not Updated");
+        }
+        logger.info("Update User ends");
         return apiResponse;
     }
